Working around Eclipse Java compiler bug
I'm not actually entirely sure who is to blame for the issue (Scala binaries or
Eclipse Java compiler), but calling `toSeq` on a Scala Map in a Java source
doesn't work.  The error reported is something along the lines of "ambiguos
method call".

While it's not ideal to workaround tooling issues, the benefits is that I/we
will be able to finally import the Play codebase in Eclipse (with Scala IDE).